pyproject.toml
[tool.poetry]
name = "ag-advisor-map"
version = "0.1.0"
description = "Agricultural advice application"
authors = [
    "Michael <heronalps@gmail.com>",
    "Brenan <lundquist.Brenan@gmail.com>",
    "Nathaniel <thenathanielwatkins@gmail.com>",
    "Eric <ericssimpson2021@gmail.com>",
    "Dawson <dawson.whipple@gmail.com>",
    "Logan <loganbugenig01@gmail.com>"
]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
fastapi = "*"
uvicorn = "*"
httpx = "*"
python-dotenv = "*"
pydantic-settings = "*"
langchain-community = "*"
langchain-openai = "*"
langchain = "*"
langgraph-cli = "*"
langgraph = "*"
openai = "*"

[tool.poetry.group.dev.dependencies]
black = "*"
flake8 = "*"
pytest = "*"
pytest-asyncio = "*"
isort = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3